Fix read qname

+++ src/read.c
@@ -91,8 +91,8 @@ int read_question(void *buffer, const int pos, struct question *q)
 int read_qname(void *buffer, const int pos, char *qname)
 {
     int current_pos = pos;
-    uint8_t length;
-    uint8_t tmp_pos = current_pos;
+    uint16_t length;
+    uint16_t tmp_pos = current_pos;
     while ((length = *(uint8_t *)(buffer + tmp_pos)))
     {
         if (((length & 0b11000000) >> 6) == 0b11)
